MG-MixF PPP-AR
In the traditional multi-frequency models, once the satellites lack a specific frequency signal, that satellite cannot contribute to the positioning solution. The mixed multi-frequency model overcomes this limitation of traditional models. However, existing PPP-AR open-source software packages are not yet user-friendly enough to handle multi-GNSS and mixed multi-frequency data due to mixed multi-frequency complex models. We have developed flexible and user-friendly open-source software for processing multi-GNSS mixed multi-frequency PPP-AR, named MG-MixF PPP-AR. In the event of a specific frequency of the triple-frequency satellite being lost, the satellite continues to participate in the process of calculating the ambiguity fixed solution. Additionally, the software batch processing of multi-day and multi-station data.